[Refactor] [Sandalphon] Use consistent value objects for bundle items

+++ judgels-backends/sandalphon/sandalphon-app/app/org/iatoki/judgels/sandalphon/problem/bundle/grading/BundleDetailResult.java
@@ -2,10 +2,10 @@ package org.iatoki.judgels.sandalphon.problem.bundle.grading;
 
 public final class BundleDetailResult {
 
-    private final long number;
+    private final int number;
     private final double score;
 
-    public BundleDetailResult(long number, double score) {
+    public BundleDetailResult(int number, double score) {
         this.number = number;
         this.score = score;
     }
